requirements.txt
pyaudio~=0.2.11
webrtcvad~=2.0.10
halo~=0.0.18
numpy==1.14.1
scipy>=1.1.0
SpeechRecognition
moviepy==1.0.3
python-vlc